[stable10] While decrypting userkey prepareEncryptionModules() must be called, keep fileid on decrypt
Below are the 2 causes addressed by this change:
1. Regression was caused because of omitting call to method
prepareEncryptionModules. This change will get the userkey encryption
work as before.

2. The shares were not available in the UI, even if they
were available in oc_share table. The reason for this
was due to the copy and rename operation done in the
decrypt-all operation. And the share table wasn't updated
with the new file. This change updates the share table
with the updated id, which helps the UI to show the share
files.

Signed-off-by: Sujith H <sharidasan@owncloud.com>